backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
asgiref==3.7.2
    # via django
django==4.2.8
    # via
    #   -r requirements.in
    #   djangorestframework
    #   drf-yasg
django-environ==0.11.2
    # via -r requirements.in
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   drf-yasg
drf-yasg==1.21.7
    # via -r requirements.in
inflection==0.5.1
    # via drf-yasg
packaging==23.2
    # via drf-yasg
psycopg[binary]==3.1.16
    # via -r requirements.in
psycopg-binary==3.1.16
    # via psycopg
pytz==2023.3.post1
    # via
    #   djangorestframework
    #   drf-yasg
pyyaml==6.0.1
    # via drf-yasg
sqlparse==0.4.4
    # via django
typing-extensions==4.9.0
    # via psycopg
uritemplate==4.1.1
    # via drf-yasg
